ci: Fail lint job on warnings
The Android Lint job will now fail if any warnings are found, in addition to errors. Previously, it only failed on errors.

This is achieved by checking the output report for "Error" or "Warning" and exiting with a failure code if either is present. The job name has also been updated to reflect this stricter check.

Additionally, the `lint-results-debug.html` and `lint-results-debug.txt` files are now included in the uploaded artifacts.